requirements.txt
aiosmtpd==1.4.2
dkimpy==1.0.5
asyncio==3.4.3
email-validator==1.3.1
